Crate calypso_repl

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

Command

Editor

Line editor

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

KeyEvent

Input key pressed and modifiers

Regex

A compiled regular expression for matching Unicode strings.

Repl

A struct for doing REPL-like activities. This does not necessarily need to fit the exact definition of REPL (Read, Eval, Print, Loop).

Enums

Cmd

Commands

Movement

Where to move

ReadlineError

The error type for Rustyline errors that can arise from I/O related errors or Errno when using the nix-rust library

Traits

Configurer

Trait for component that holds a Config.

Type Definitions

Eval

A closure that evaluates some input with some context type, and returns either Some(String) or None. None indicates to the REPL handler that it should break the loop.